+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `serviceData` in `BleDevice`
 * Add `AndroidScanMode` and `reportDelayMillis` to `AndroidOptions` for scanning
 * Add support for BLE scanning from background services (ForegroundTask) on Android. `PermissionHandler` is now activity-aware and succeeds silently if permissions are already granted when no activity is available
+* Always use custom scan filter on Windows
 
 ## 1.1.0
 * Add readRssi method